# backend/src/Apps/Collections/openapi.yaml
openapi: 3.0.1
info:
  title: Dummy API (mojeEUC)
  description: Dummy service (mojeEUC) interface specification.
  termsOfService: https://cyberfox.cz/
  contact:
    name: Petr Tichý
    url: https://cyberfox.cz/
    email: petr.tichy@cyberfox.cz
  license:
    name: CyberFox Closed Source Commercial License (CFCSCL)
    url: https://cyberfox.cz/
  version: v1
servers:
  - url: https://cf-meuc-dev.azure-api.net/dummy/v1
paths:
  /author:
    post:
      tags:
        - Author
      summary: Creates a new entity.
      description: Creates a new entity.
      operationId: AuthorCreate
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthorCreateRequest'
            example:
              firstName: string
              lastName: string
      responses:
        '201':
          description: Author was created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthorResponse'
              example:
                firstName: string
                lastName: string
                isActive: true
                id: string
        '400':
          description: Author creation request is invalid.
        '401':
          description: There was en error authentication user.
        '403':
          description: User is not authorized to perform requested operation.
        '500':
          description: There was a general error during request processing.
        '501':
          description: Requested operation is not implemented.
    get:
      tags:
        - Author
      summary: Retrieves a list of authors based on the specified criteria.
      description: Retrieves a list of authors based on the specified criteria.
      operationId: AuthorList
      parameters:
        - name: onlyActive
          in: query
          description: Limits the resulting list to only active authors.
          schema:
            type: boolean
            default: true
        - name: if-Modified-Since
          in: header
          description: Format - date-time (as date-time in RFC3339). Limits the resulting list to only authors that have been modified since.
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
              example:
                - string
        '304':
          description: Any author were modified since requested date.
        '401':
          description: There was en error authentication user.
        '403':
          description: User is not authorized to perform requested operation.
        '500':
          description: There was a general error during request processing.
        '501':
          description: Requested operation is not implemented.
  '/author/{id}':
    delete:
      tags:
        - Author
      summary: Deletes entity by ID.
      description: Deletes entity by ID.
      operationId: AuthorDelete
      parameters:
        - name: id
          in: path
          description: ''
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Author was deleted successfully.
        '401':
          description: There was en error authentication user.
        '403':
          description: User is not authorized to perform requested operation.
        '404':
          description: Author was not found.
        '409':
          description: Author is not active.
        '412':
          description: Author was modified by another request.
        '500':
          description: There was a general error during request processing.
        '501':
          description: Requested operation is not implemented.
    get:
      tags:
        - Author
      summary: Gets entity by ID.
      description: Gets entity by ID.
      operationId: AuthorGet
      parameters:
        - name: id
          in: path
          description: ''
          required: true
          schema:
            type: string
        - name: onlyActive
          in: query
          description: ''
          schema:
            type: boolean
            default: true
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthorResponse'
              example:
                firstName: string
                lastName: string
                isActive: true
                id: string
        '304':
          description: Author was not modified.
        '401':
          description: There was en error authentication user.
        '403':
          description: User is not authorized to perform requested operation.
        '404':
          description: Author was not found.
        '409':
          description: Author is not active (if requested onlyActive record).
        '500':
          description: There was a general error during request processing.
        '501':
          description: Requested operation is not implemented.
    put:
      tags:
        - Author
      summary: Returns a list of entities.
      description: Returns a list of entities.
      operationId: AuthorUpdate
      parameters:
        - name: id
          in: path
          description: ''
          required: true
          schema:
            type: string
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthorUpdateRequest'
            example:
              firstName: string
              lastName: string
      responses:
        '200':
          description: Author was updated successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthorResponse'
              example:
                firstName: string
                lastName: string
                isActive: true
                id: string
        '400':
          description: Author update request is invalid.
        '401':
          description: There was en error authentication user.
        '403':
          description: User is not authorized to perform requested operation.
        '404':
          description: Author was not found.
        '409':
          description: Author is not active.
        '412':
          description: Author was modified by another request.
        '500':
          description: There was a general error during request processing.
        '501':
          description: Requested operation is not implemented.
...
components:
  schemas:
    AuthorCreateRequest:
      required:
        - firstName
        - lastName
      type: object
      properties:
        firstName:
          minLength: 1
          type: string
          description: The first name of the author.
        lastName:
          minLength: 1
          type: string
          description: The last name of the author.
      additionalProperties: false
      description: Request to create an author.
    AuthorResponse:
      required:
        - firstName
        - id
        - isActive
        - lastName
      type: object
      properties:
        firstName:
          minLength: 1
          type: string
          description: The first name of the author.
        lastName:
          minLength: 1
          type: string
          description: The last name of the author.
        isActive:
          type: boolean
          description: Indicates whether the author is active (or soft-deleted).
        id:
          minLength: 1
          type: string
          description: Author ID.
      additionalProperties: false
      description: General response to author requests.
    AuthorUpdateRequest:
      required:
        - firstName
        - lastName
      type: object
      properties:
        firstName:
          minLength: 1
          type: string
          description: The first name of the author.
        lastName:
          minLength: 1
          type: string
          description: The last name of the author.
      additionalProperties: false
      description: Request to update an author.
...
  securitySchemes:
    apiKeyHeader:
      type: apiKey
      name: Ocp-Apim-Subscription-Key
      in: header
    apiKeyQuery:
      type: apiKey
      name: subscription-key
      in: query
security:
  - apiKeyHeader: [ ]
  - apiKeyQuery: [ ]